+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 * #4224: oauth2_proxy is now hosted by a new organisation (https://oauth2-proxy.github.io/oauth2-proxy/)
 * #4299: Bump vertx from 3.8.3 to 3.9.0
 
+## 0.31.4
+*  #4704: [OpenShift 3.11] Console returns 500 internal error when configured with custom certificate
+
 ## 0.31.3
 * #4650: Install uncaught exception handler in address-space-controller and standard-controller